changed attribute to integer

+++ HSPR_AZ_hpc.py
@@ -58,13 +58,13 @@ description
     Trancription leakage (default: 0.01)
 
 --hilHalfSaturation,-hhs
-    The conc of inducer/repressor at half-max transcriptional rate (default: 1.0)
+    The conc of inducer/repressor at half-max transcriptional rate (default: 1)
 
 --maxA1,-ma1
-    a1, the max transcription rate of A1 (default: 10.0)
+    a1, the max transcription rate of A1 (default: 10)
 
 --foldedProduction,-fpp
-    a7, folded protein production rate (default: 10.0)
+    a7, folded protein production rate (default: 10)
 
 --outputFormat,-ofm
     Whether to save Gillespi simulation output as csv or picklefile (default: csv)
@@ -167,7 +167,7 @@ def param_spec(opt):
         'init_HSFB': 1,
         'Time': 0.0,
         ## Maximum expression level in Hill equation
-        'a1': float(opt.ma1),
+        'a1': int(opt.ma1),
         'a2': 100.0,
         'a3': 5.0,
         'a4': 5.0,
@@ -176,12 +176,12 @@ def param_spec(opt):
         'a7': int(opt.fpp), #folded protein production rate
         'a8': 5.0,
         ## Ka in Hill equation
-        'h1': float(opt.hhs),
-        'h2': float(opt.hhs),
-        'h3': float(opt.hhs),
-        'h4': float(opt.hhs),
-        'h5': float(opt.hhs),
-        'h6': float(opt.hhs),
+        'h1': int(opt.hhs),
+        'h2': int(opt.hhs),
+        'h3': int(opt.hhs),
+        'h4': int(opt.hhs),
+        'h5': int(opt.hhs),
+        'h6': int(opt.hhs),
         ## association rates
         'c1': 10.0, #between A1 and HSPR
         'c3': float(opt.amh), #between MMP and HSPR
